Move TextToImageInferenceOutput to a shared location.
