--- shortfin/python/shortfin_apps/sd/components/TextToImageInferenceOutput.py
@@ -1,14 +1,0 @@
-# Copyright 2025 Advanced Micro Devices, Inc.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-from PIL.Image import Image
-
-from dataclasses import dataclass
-
-
-@dataclass
-class TextToImageInferenceOutput:
-    image: Image
