+++ shortfin/python/shortfin_apps/sd/components/generate.py
@@ -19,6 +19,7 @@ from shortfin_apps.types.Base64CharacterEncodedByteSequence import (
 )
 
 from shortfin_apps.utilities.image import png_from
+from shortfin_apps.types.TextToImageInferenceOutput import TextToImageInferenceOutput
 
 import shortfin as sf
 
@@ -29,7 +30,6 @@ from .io_struct import GenerateReqInput
 from .messages import SDXLInferenceExecRequest
 from .service import SDXLGenerateService
 from .metrics import measure
-from .TextToImageInferenceOutput import TextToImageInferenceOutput
 
 logger = logging.getLogger("shortfin-sd.generate")
