+++ shortfin/python/shortfin_apps/types/TextToImageInferenceOutput.py
